Use cyan to highlight the current word
The previous behavior was to use the same color for the current word as
for all the other letters in the grid. Using this method, I often found
myself getting lost and hunting to find my position in the grid.
Changing the current word to a different color was a big help.

This is a quick-and-dirty solution that suits my needs. A proper
solution would probably involve adding support for user-defined color
themes.