Simplify navigation when using the arrow keys
The new behavior of the arrow keys is to simply move the cursor in the
corresponding direction. Previously, the behavior depended on whether
the current word went across or down. For example, supposed I press the
down arrow. If the current word went across, the down arrow would toggle
the current word to go down. If the current word went down, the down
arrow would move the cursor to the next letter in the word.