Stop wrapping at grid edges
Suppose the cursor is at the far right column of the grid and we take
some action that would ordinarily move the cursor to the right (such as
pressing the right arrow key). The old behavior would move the cursor to
the far left column on the next row. I think that would be more often
annoying than useful for me, so this commit changes the behavior so that
the cursor doesn't move in this situation.