Use overwrite mode when typing letters
Suppose the current word has a few letters filled in already and looks
like `_UZZ__` with the cursor in the first cell. Previously, typing `P`
would insert a `P` in the first cell, and the cursor would skip over the
already-filled cells, ending up in the fifth cell. This commit changes
the behavior so that the cursor doesn't skip the already-filled cells,
meaning it ends up in the `U` cell.

I find this way of entering letters to be more natural. I prefer to just
type the whole word, even if some letters are already present, than to
keep track of which letters are being skipped as I'm typing.