Don't jump to the next word after typing to the end of the current word
When I'm completing crosswords, I often like to work within a "block",
so when I type in a whole word, I'll switch from across to down and work
on the words that cross the one I just filled in. That's more cumbersome
to do using the previous behavior, where the cursor would jump to the
next across word after I filled one in, because the next across word is
almost always in a different block. Then I have to navigate my way back
to the current block. With this change, the cursor remains in the word I
just filled in, so it's still in the same block.

This also includes the reverse behavior of not skipping to the previous
word when reaching the beginning of the current word, although I'm not
sure of the use case that would make that happen. The code reaches that
point when the "Back Tab" key is pressed, but I don't know what produces
a Back Tab.