Treat the Delete key more like Backspace
Currently, the app can't distinguish between Delete and Backspace. The
previous behavior was a bit of a mix between what I'd expect out of
Delete and Backspace: It would clear the current cell (like Delete) and
move the cursor to the previous cell (like Backspace). I think I'm more
likely to prefer a Backspace-style behavior, so this commit implements
that. Now, the current cell is left untouched, the cursor moves to the
previous cell, and the previous cell is cleared.