Make Space key clear the current cell
I think of the Space key as more of a blank character than as a control
character, so it's more intuitive to me to enter a blank than to toggle
the clue direction. Using Space in this way is helpful when trying to
remove an incorrect letter when you don't know what the correct
replacement letter should be yet.