Upgrade the blessed dependency
The purpose of this upgrade is to pick up blessed's recently-added
support for receiving separate key events for the Backspace and Delete
keys.

An unrelated change in blessed is that the `dim` method is no longer
supported. (The docs say it's deprecated, but I get an error when
calling it.) My terminal doesn't support `dim` anyway, so I simply
removed all the calls to `dim` in the code.