Use parens to encircle letters
Sometimes a puzzle circles letters, using the circled letters as an
additional twist. A circled letter had been represented on the board by
a separate Unicode symbol showing the letter inside a circle. My
combination of terminal and fonts had trouble rendering these Unicode
symbols. To get around this, I added parentheses around the letter
instead.

The parentheses go in the "markup" area in the cell. The markup area is
a single character before and after the actual letter. Usually it's
filled with a space. Previously, a colon was added to this area to
indicated a revealed letter, and a dot was added to indicate a corrected
letter. I had to remove these to accommodate the parentheses. This is OK
for me, though, because I don't really care about revealed/corrected
letters.